+++ src/routes/authentication.js
@@ -5,8 +5,42 @@ import * as authentication from '../controllers/authentication';
 
 const authenticationRoutes = Router();
 
-authenticationRoutes.route('/authentication/signup').post(authentication.signup);
-authenticationRoutes.route('/authentication/signin').post(authentication.signin);
+authenticationRoutes.route('/auth/signup').post(authentication.signup);
+authenticationRoutes.route('/auth/signin').post(authentication.signin);
+
+// Facebook authentication
+authenticationRoutes.route('/auth/facebook').get(authentication.socialAuth('facebook', {
+  scope: ['email'],
+}));
+authenticationRoutes.route('/auth/facebook/callback').get(authentication.socialAuthCallback('facebook'));
+
+// Google authentication
+authenticationRoutes.route('/auth/google').get(authentication.socialAuth('google', {
+  scope: [
+    'https://www.googleapis.com/auth/userinfo.profile',
+    'https://www.googleapis.com/auth/userinfo.email',
+  ],
+}));
+authenticationRoutes.route('/auth/google/callback').get(authentication.socialAuthCallback('google'));
+
+// Linkedin authentication
+authenticationRoutes.route('/auth/linkedin').get(authentication.socialAuth('linkedin', {
+  scope: [
+    'r_basicprofile',
+    'r_emailaddress',
+  ],
+}));
+authenticationRoutes.route('/auth/linkedin/callback').get(authentication.socialAuthCallback('linkedin'));
+
+// Twitter authentication
+authenticationRoutes.route('/auth/twitter').get(authentication.socialAuth('twitter'));
+authenticationRoutes.route('/auth/twitter/callback').get(authentication.socialAuthCallback('twitter'));
+
+// Github authentication
+authenticationRoutes.route('/auth/github').get(authentication.socialAuth('github', {
+  scope: 'user:email',
+}));
+authenticationRoutes.route('/auth/github/callback').get(authentication.socialAuthCallback('github'));
 
 // authentication middleware
 authenticationRoutes.use((req, res, next) => {
@@ -21,7 +55,7 @@ authenticationRoutes.use((req, res, next) => {
     // verifies secret and checks exp
     jwt.verify(token, env.jwtSecret, (err, decoded) => {
       if (err) {
-        return res.json({ success: false, message: 'Authentication failed.' });
+        return res.status(403).send({ success: false, message: 'Authentication failed.' });
       }
       // if everything is good, save to request for use in other routes
       req.decoded = decoded;